refactor: Moved audit extensions.
