+++ src/Chrono/Entities/Common/BaseAuditableEntity.cs
@@ -7,17 +7,3 @@ public abstract class BaseAuditableEntity : BaseEntity
     public DateTime LastModified { get; set; }
     public User LastModifiedBy { get; set; }
 }
-
-public static class BaseAuditableEntityExtensions
-{
-    public static bool IsPermitted(this BaseAuditableEntity context, string userId)
-    {
-        if (string.IsNullOrEmpty(userId))
-        {
-            return false;
-        }
-
-        // Only the owner is permitted to view/edit the object.
-        return context.CreatedBy.UserId == userId;
-    }
-}
