+++ src/Chrono/Features/Categories/GetCategories.cs
@@ -1,6 +1,6 @@
 using Chrono.Common.Api;
 using Chrono.Common.Interfaces;
-using Chrono.Entities.Common;
+using Chrono.Features.Audit;
 using Chrono.Features.Users;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
