+++ src/Chrono/Features/TaskLists/UpdateTaskList.cs
@@ -2,7 +2,7 @@
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
 using Chrono.Entities;
-using Chrono.Entities.Common;
+using Chrono.Features.Audit;
 using Chrono.Features.Users;
 using FluentValidation;
 using MediatR;
