+++ src/Chrono/Features/Tasks/CreateTask.cs
@@ -2,7 +2,7 @@ using Chrono.Common.Api;
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
 using Chrono.Entities;
-using Chrono.Entities.Common;
+using Chrono.Features.Audit;
 using Chrono.Features.Categories;
 using Chrono.Features.TaskLists;
 using Chrono.Features.Users;
