+++ src/Chrono/Features/Tasks/GetTask.cs
@@ -1,7 +1,7 @@
 using Chrono.Common.Api;
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
-using Chrono.Entities.Common;
+using Chrono.Features.Audit;
 using Chrono.Features.Users;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
