+++ src/Chrono/Features/Users/UpdateUserSettings.cs
@@ -2,7 +2,7 @@
 using Chrono.Common.Exceptions;
 using Chrono.Common.Interfaces;
 using Chrono.Entities;
-using Chrono.Entities.Common;
+using Chrono.Features.Audit;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
